[peerdist] Reduce number of concurrent requests to origin server
Origin server requests are likely to be significantly more expensive
than local peer requests: they consume sparse uplink bandwidth and, if
using HTTPS, substantial local resources.

Start with a low number of concurrent block downloads and adjust this
number in response to the availability of local peers, in order to
avoid hammering the origin server with the full concurrency limit.

Signed-off-by: Michael Brown <mcb30@ipxe.org>