[netdevice] Allow duplicate MAC addresses
Many laptops now include the ability to specify a "system-specific MAC
address" (also known as "pass-through MAC"), which is supposed to be
used for both the onboard NIC and for any attached docking station or
other USB NIC.  This is intended to simplify interoperability with
software or hardware that relies on a MAC address to recognise an
individual machine: for example, a deployment server may associate the
MAC address with a particular operating system image to be deployed.

This therefore creates legitimate situations in which duplicate MAC
addresses may exist within the same system.  Remove the code that
rejects attempts to register a network device with a duplicate MAC.

Signed-off-by: Michael Brown <mcb30@ipxe.org>